+++ src/main/java/com/nhnacademy/bookstore/orderset/order/entity/Order.java
@@ -12,8 +12,8 @@ package com.nhnacademy.bookstore.orderset.order.entity;
 import com.nhnacademy.bookstore.coupon.entity.member.MemberCoupon;
 import com.nhnacademy.bookstore.orderset.order.dto.request.OrderPostRequest;
 import com.nhnacademy.bookstore.orderset.order.dto.request.OrderRequest;
-import com.nhnacademy.bookstore.orderset.order_detail.entity.OrderDetail;
-import com.nhnacademy.bookstore.orderset.order_state.entity.OrderState;
+import com.nhnacademy.bookstore.orderset.orderdetail.entity.OrderDetail;
+import com.nhnacademy.bookstore.orderset.orderstate.entity.OrderState;
 import com.nhnacademy.bookstore.user.customer.entity.Customer;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
