+++ src/main/java/com/nhnacademy/bookstore/orderset/order/service/OrderService.java
@@ -24,10 +24,10 @@ import com.nhnacademy.bookstore.orderset.order.dto.request.OrderRequest;
 import com.nhnacademy.bookstore.orderset.order.entity.Order;
 import com.nhnacademy.bookstore.orderset.order.mapper.OrderMapper;
 import com.nhnacademy.bookstore.orderset.order.repository.OrderRepository;
-import com.nhnacademy.bookstore.orderset.order_detail.entity.OrderDetail;
-import com.nhnacademy.bookstore.orderset.order_detail.repository.OrderDetailRepository;
-import com.nhnacademy.bookstore.orderset.order_state.entity.OrderState;
-import com.nhnacademy.bookstore.orderset.order_state.service.OrderStateService;
+import com.nhnacademy.bookstore.orderset.orderdetail.entity.OrderDetail;
+import com.nhnacademy.bookstore.orderset.orderdetail.repository.OrderDetailRepository;
+import com.nhnacademy.bookstore.orderset.orderstate.entity.OrderState;
+import com.nhnacademy.bookstore.orderset.orderstate.service.OrderStateService;
 import com.nhnacademy.bookstore.paymentset.payment_history.entity.PaymentHistory;
 import com.nhnacademy.bookstore.paymentset.payment_history.repository.PaymentHistoryRepository;
 import com.nhnacademy.bookstore.paymentset.payment_method.entity.PaymentMethod;
