--- src/main/java/com/nhnacademy/bookstore/orderset/order_state/converter/OrderStateTypeConverter.java
@@ -1,26 +1,0 @@
-package com.nhnacademy.bookstore.orderset.order_state.converter;
-
-import com.nhnacademy.bookstore.orderset.order_state.entity.vo.OrderStateType;
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-
-@Converter(autoApply = true)
-public class OrderStateTypeConverter implements AttributeConverter<OrderStateType, String> {
-    @Override
-    public String convertToDatabaseColumn(OrderStateType orderStateType) {
-        return orderStateType.getDescription();
-    }
-
-    @Override
-    public OrderStateType convertToEntityAttribute(String dbData) {
-        if (dbData == null || dbData.isEmpty()) {
-            return null;
-        }
-        for (OrderStateType type : OrderStateType.values()) {
-            if (type.getDescription().equals(dbData)) {
-                return type;
-            }
-        }
-        throw new IllegalArgumentException("Unknown value: " + dbData);
-    }
-}
