+++ src/main/java/com/nhnacademy/bookstore/review/service/ReviewServiceImpl.java
@@ -14,8 +14,8 @@ import com.nhnacademy.bookstore.imageset.entity.Image;
 import com.nhnacademy.bookstore.imageset.entity.ReviewImage;
 import com.nhnacademy.bookstore.imageset.repository.ImageRepository;
 import com.nhnacademy.bookstore.imageset.repository.ReviewImageRepository;
-import com.nhnacademy.bookstore.orderset.order_detail.entity.OrderDetail;
-import com.nhnacademy.bookstore.orderset.order_detail.repository.OrderDetailRepository;
+import com.nhnacademy.bookstore.orderset.orderdetail.entity.OrderDetail;
+import com.nhnacademy.bookstore.orderset.orderdetail.repository.OrderDetailRepository;
 import com.nhnacademy.bookstore.point.dto.request.ReviewPointAwardRequest;
 import com.nhnacademy.bookstore.point.service.PointService;
 import com.nhnacademy.bookstore.review.dto.request.ReviewCreateRequestDto;
